+++ test/aoj/dpl/5_E.test.cpp
@@ -1,4 +1,4 @@
-#define PROBLEM "https://onlinejudge.u-aizu.ac.jp/courses/library/7/DPL/5/DPL_5_F"
+#define PROBLEM "https://onlinejudge.u-aizu.ac.jp/courses/library/7/DPL/5/DPL_5_E"
 
 #include <bits/stdc++.h>
 using namespace std;
@@ -507,7 +507,7 @@ namespace atcoder {
 // modint.hpp end here
 
 
-#include "../../../../../math/mod-factorial.hpp"
+#include "../../../math/mod-factorial.hpp"
 using namespace math;
 
 int main() {
@@ -515,5 +515,6 @@ int main() {
   cin >> n >> k;
   ModFactorial mf;
 
-  cout << mf.combination(n - 1, k - 1).val() << endl;
+  cout << mf.combination(k, n).val() << endl;
 }
+
